// Front-Amazon/src/interfaces/product.ts
export interface Product {
disabled: boolean;
  quantity: number;
  product: any;
  productId: number;
  title:string
    name: string;
    price: number;
    stock:number;
    weight:number;
    description: string;
    category: string;
    image: string;
    supplierId:number;
    subCategoryId:number;
}